improved fullscreen handling
